Assume no funny business
As I write the solution to reverse a String in C#, I'm finding that it's not so easy. Well, it's actually simply, but the code points are handled in some languages can be a nightmare. As a result, I'm assuming one language character maps to one visual character for these algorithms.