c hello world: use puts in place of printf
The original hello world program for C uses the `printf` function to
print a constant string to the screen. However, there's two issues with
that:

1. The "f" in `printf` stands for "formatted", so it expects some format
   sequences in its first argument. Granted, this is a little nitpicky
   and only really makes a semantic difference.
2. It doesn't print a newline after the string. This is against common
   convention (see, for example, bash's `echo), and is pretty ugly with
   shells that display it (zsh adds a trailing % in this case).

The `puts` function is designed to print a constant string (that is,
without format sequences) to stdout and add a newline character at the
end, so it's a much better fit for this application.