string reverse in C
Add a naive string reverse implementation in C.
It reads stdin line by line and prints the reverse version of that
string to stdout. The behaviour is quite similar to that of the
widely used Linux utility `rev`.

Note that this is a relatively trivial implementation, so it only
reverses each character from the input string, each character being 1
Byte wide. This means that things like accents, Chinese characters, and
emojis won't work.