edit hello world in brainfuck
The previous example for Hello World in Brainfuck was a great solution in
terms of compactness, though it had a few shortcomings which I want to address
with this edit.

The main problem is readability. As the previous example was probably
generated using a text generator, it doesn't make much sense if you try
to break it down.

Also, the solution relies on the fact that the cell pointer wraps around
if you decrement it past zero, which isn't the case in some
implementations.

This version addresses both of those issues, albeit at the cost of
compactness (it is, after all, twice as long).