Game of Life in C (#357)
An implementation of Conway's Game of Life in C.

The code tries to be as standard-agnostic as possible, but due to the nature of the program (considering graphical output and such) it was basically impossible to make it completely portable. A Decent Shell™ and a POSIX compliant operating system are definitely required (which doesn't currently apply to `cmd` under Windows).