New language: Vimscript
Vimscript is an old and weird language, and its sole purpose is to
script the vim editor. There is barely any documentation on it, and as a
general rule of thumb, it always behaves as unpredictably as it possibly
can.

That said, it is obviously quite fun to write sample programs in it, so
hereby include it in this repository.