Hello World in Vimscript
Simple "Hello World" implementation in Vimscript. To run this solution,
use:

    $ vim -u hello-world.vim

The autocmd will automatically execute the function upon starting the
editor.

Even though an echo command exists in Vimscript, I opted to append the
text to the buffer instead. This is because the echo command only allows
one line to be visible at a time, and later solutions (such as FizzBuzz)
might not exactly benefit from this limitation.