Fizz Buzz in Vimscript
Kind of an ugly solution, but hey, it's Vimscript.

As with the Hello World and Reverse String solutions, this appends the
resulting text to the buffer (and moves to the top at the end so the
result is actually visible). See commit 6229cffe0371 ("Hello World in
Vimscript") for rationale.

Execute using:

    $ vim -u fizz-buzz.vim

Again, an autocommand is used to execute the function automatically upon
entering the buffer.