New Language: TeX
Okay, okay. This is kind of stretching the definition of a "programming
language". Arguably (La)TeX is almost exclusively used for text
formatting and typesetting. In my defense though, it *is* technically
turing complete, and some people are even crazy enough to write rather
complex programs with it [1].

I haven't tried, but I figure at least FizzBuzz and a Quine should be
easily doable in TeX.

Anyways, add TeX to the collection.

[1] https://ctan.org/pkg/basix