Add debut date for vimscript
After some digging around in various dusty corners of the internet, I've
managed to find a relatively definitive source for the release date of
vim's scripting language.

In the release archive (ftp://ftp.vim.org/pub/vim/unix/) there's a list
of really old vim versions readily available for download. I've dug
through these and noticed that back in the day, included with the source
tarball came a handy text document describing the differences between each
major version.

As it happens, the document for version 5.0 of vim includes the
following:

> Built-in script language                                   *new-script*
> ------------------------
>
> With a few extra commands and an expression evaluator, simple but
> powerful scripts can be written.  Commands include ":if" and ":while".
> Expressions can be used to manipulate numbers and strings.  The '='
> register can be used to directly insert the result of an expression.
> See |expression|.

The head of the document also states that the last modification happened
on February 19th, 1998. Thus, we can reasonably conclude that the debut of
vimscript happened somewhere in 1998.