Fizz Buzz in TeX
This doesn't exactly follow the usual "algorithm" used to implement Fizz
Buzz. Rationale: plain TeX is a seriously crippled language which is
primarily (only?) used for typesetting and therefore doesn't support
even the most basic mathematical operations, such as modulo.
Note that one could technically use a package to accomplish this, but
what fun is that?

So, this implementation uses a different approach. Basically it uses two
counters to count up to three and five respectively. When the "three"
counter is 3, it prints "Fizz", when the "five" counter is 5, it prints
"Buzz" (note that this automatically covers the "FizzBuzz" case). If
neither of the strings were printed, it sets a flag to print the
counter.

This uses only plain TeX primitives (from <=1990!), and I'm oddly proud
of that fact.