Shell Sort .cpp
The shell sorting technique is based on the insertion sort. In the insertion sort sometimes we need to shift large blocks to insert an item in the correct location. Using shell sort, we can avoid a large number of shifting. The sorting is done with a specific interval. After each pass, the interval is reduced to make a smaller interval.