Insertion Sort Objective-C. m
The idea behind insertion sort is to take 2 for loops and run the outer loop starting at 1 and increment it one at a time through a mutable data set. On each of these iterations then, the inner loop then checks each of the values at the mutable array's index to see if values need to be reordered.