Create duplicate_character_counter.rb
code returns  count of all duplicate, case-sensitive, alphanumeric characters, that occur more than once in the given string. Characters should also be presented in the order in which they appear in the string. For instance, 'abba' should output 'a' as the first character that has a duplicate.

Error message is Usage: please provide a string